# Cache item info for crafting GUI full-text search