#### Summary
Performance "Cache item info for crafting GUI full-text search"

#### Purpose of change
- Fixes part of #63091 namely: `Cache results of "d:some item description"`. But it is done differently and better.

Getting info on an item `item.info( true )` made more than 95% of the searching time (1m11s on my PC). When cached (rerun), the search takes about 400 ms! Making it over 100x faster.

Speeds up on multiple levels:
1. If you cancel the full-text search, it still cached the data it went through.
   - Full-text search for something. Realize you made a mistake while it searches. ESC, fix your prompt and do a full-text search again. Essentially all the time from the previous (cancelled) search is reused!
3. If you have multiple full-text searches in a prompt, the first caches all the data and the next ones use that cache.
4. Refining your search prompt is super fast now. Try one (takes a minute), see results, try another one (takes like 400ms).
5. You can search; close the crafting GUI to look at what you are wearing etc (as long as you don't take a turn); reopen the crafting GUI; the cache still works, so search it again!
6. I noticed switching a crafter when a search prompt was used, it re-searches it. With this, it takes no time at all! Btw. all searches are avatar character-centric, so they are agnostic to the crafter Character.

#### Describe the solution

- Cache the items in ~`src/uistate.h`. Should the code live somewhere else?~ a variable local to `src/recipe_dictionary.cpp`.
- In full-text search, use the cache function instead of the regular function.
- Cache only for this turn.

#### Describe alternatives you've considered

Making the regular function cache automatically. It would cause more changes to test.

The item info probably breaks (is not up to date) when you do something in less than a turn. I considered fixing that, but first I would need to find an example. Maybe it will matter more if it is used someplace else, like crafting GUI in general. And if it gets used.

#### Testing

Before
1. Searched before this PR. It takes ages: `d:close` takes 1min8sec.

After
1. Searched now. It takes ages. Actually, it takes less time `d:close` takes 56sec now. The cache is utilized even for the first full-text search apparently. It has 845 hits and 3318 misses with a clear cache. The second time, it has all hits and no misses.
2. Cancel the search and redo it - it picks up (caching) where it left off.
3. Same if you let the search finish.
4. Multiple full-text searches, like the example `d:close to skin,d:HEAD`. Took 55 seconds. Nice bonus: the percentage progress now concludes at 100% and finishes, before, it made the second 0 to 100% search for the second full-text search.
5. Search, close crafting GUI, open it, and search. The second search is fast.
6. Switching crafter with a full-text prompt is fast now. (And still respects item inventories etc. It seems to work!)
7. Waiting for a turn invalidates the cache.

Am I respecting item variants? Or something? Does it break for something?
#### Additional context

This could be used more. Like connected to the crafting GUI. I wanted to start with a super simple and useful implementation. We can do the rest in subsequent PRs.

The text might need to be wrapped for other use cases. In searching there is no wrapping.