@Brambor 2024-11-17
Test failure unrelated:
```
(~[slow] ~[.],starting_items)=> (continued from above) ERROR : src/newcharacter.cpp:685 [void Character::add_profession_items()] Failed to place 3 items in inventory for profession Crossbow Hunter
```
https://github.com/CleverRaven/Cataclysm-DDA/actions/runs/11873797927/job/33089014493?pr=77914

 - Probably related to one off the NPC through avatar creation PRs: #76935