@Brambor 2024-12-26
The cache survives save-load cycle.

Potentional bugs:

**World bug**
Make worlds A and B. Get them to the same turn. Example: copy-paste a world.

Fulltext searching items in world A will cache the results. Load world B. It will use this cache.

It is unlikely that anyone will ever encounter this bug.

**Avatar bug**
Use fulltext search to cache info.

Switch conciusenes from avatar A to another character B (I assume this does not take a turn, if it does, cache is correctly invalidated). Also, I assume this can be done with choosing a new leader. Debug menu-only bugs that are solved by waiting a turn are not worth fixing.

Use cached info with avatar B.